requirements.txt
# Web Server
fastapi
uvicorn
python-multipart
requests

# AI & LLM (Google GenAI Official SDK)
google-genai
langchain-text-splitters
pypdf
Pillow
python-dotenv

# Database & Storage
supabase
redis/upstash-redis

# Environment, Schema & Security
pydantic
PyJWT
cryptography
Pillow
requests